+++ R/calmr-package.R
@@ -1,0 +1,10 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @import data.table
+#' @importFrom lifecycle deprecated
+#' @importFrom rlang .data
+## usethis namespace: end
+
+NULL